+++ RELEASE-NOTES.md
@@ -1,5 +1,6 @@
 ## 0.9: 
 
+* #7 Log file rotation configuration
 * #8  Switch to ServiceLoader-based polymorphic configs for AppenderFactory
 * #9  Upgrade to Bootique 0.15
 * #10 'log.useLogbackConfig' : support for Logback XML file.